<docs>(networking_cs): Add final notes on layer 3
Added the final notes on layer 3, I removed information on traceroute as
I feel this is beyond what I need to know and can also google this
information at a later time.

For me understand just what protocol it uses is good enough and most
vital at this stage for me.